fix(aria): stop console header overflow; move knowledge count to Sessions row
The Console/Swarms tabs + brain + dock + new-chat overflowed the 242px panel and
clipped (Dock -> 'Doc'). Dock and New Chat are now square icon buttons, and the
knowledge count moves to the Sessions header next to refresh where it reads as a
project stat instead of crowding the tab row.
